pyproject.toml
[tool.poetry]
name = "auto_repair_shop"
version = "0.1.0"
description = ""
authors = ["Кристина Ратник <kristi.2001@yandex.ru>", "Константин Григаревичюс <krgrigarevichyus@stud.etu.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.4"
psycopg2 = "^2.9.5"
Pillow = "^9.3.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
